de acuerdo a feedback se dividio archivo auth de firestore, se testeo la funcion logout
